+++ client/src/utils/days.ts
@@ -1,0 +1,3 @@
+export function daysInMonth(year: number, month: number): number {
+  return new Date(year, month, 0).getDate();
+}
