+++ client/src/utils/index.ts
@@ -1,0 +1,3 @@
+export function range(size: number, start?: number): number[] {
+  return [...Array(size).keys()].map((i) => i + (start || 0));
+}
